Apache configuration syntax highlighting for GtkSourceView
This file was originally written in 2010 and updated in 2010-2011.
A bug report was opened to include it officially in GtkSourceView.
See <https://bugzilla.gnome.org/show_bug.cgi?id=612368>.
Although it was close to being included, it was never committed.
This version corresponds to the last one proposed in that bug report.